изменила на инт

+++ app/src/main/java/otus/homework/customview/CategoryOverallSpending.kt
@@ -6,5 +6,5 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class CategoryOverallSpending(
     val category: Category,
-    val amount: Float
+    val amount: Int
 ) : Parcelable
